requirements.txt
flask==3.0.3
gunicorn==22.0.0
scikit-learn==1.4.2
numpy==1.26.4
joblib==1.4.2
python-dotenv==1.0.1
openpyxl==3.1.2
beautifulsoup4==4.12.3
requests==2.31.0
python-docx==1.1.2
markupsafe==2.1.5
